+++ src/lib/components/toolbar/base-toolbar/terra-base-toolbar.component.ts
@@ -10,19 +10,8 @@ import { Component, Input } from '@angular/core';
 /** @deprecated since v5.0. Please use mat-toolbar instead */
 export class TerraBaseToolbarComponent {
     /**
-     * @deprecated since v4. Breadcrumbs have its own styles now.
-     */
-    @Input()
-    public inputIsBreadcrumbs: boolean;
-
-    /**
      * @description makes toolbar stick to the top while scrolling
      */
     @Input()
-    public isSticky: boolean;
-
-    constructor() {
-        this.isSticky = false;
-        this.inputIsBreadcrumbs = false;
-    }
+    public isSticky: boolean = false;
 }